Separate tests for stopword and tokenizer from hasher's one
